requirements.txt
requests==2.23.0
aiohttp==3.6.2
beautifulsoup4==4.9.1
bs4==0.0.1
google-api-python-client==1.8.4
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
oauth2client==4.1.3
PageRange==0.4
pytype==2020.2.6
PyYAML==5.3.1
selenium==3.141.0
tenacity==6.2.0
tqdm==4.46.0
